+++ examples/create_validation_request_anchor.rs
@@ -4,25 +4,25 @@
 //! cargo run --example create_validation_request_anchor -- --node http://localhost:20100 --account 3nhMYfA59MWaxBRjfHPKSYH9S4W5HdZZ721jozVdeToBGvXTU8.export
 use anyhow::Context as AnyhowContext;
 use clap::AppSettings;
+use concordium_base::web3id::v1::anchor::{
+    CredentialType, IdentityProviderMethod, RequestedIdentitySubjectClaims,
+    UnfilledContextInformation, VerificationRequestData,
+};
+use concordium_base::web3id::Web3IdAttribute;
 use concordium_base::{
     common::cbor,
     id::id_proof_types::{AtomicStatement, AttributeInRangeStatement},
     web3id::{
         did::Network,
-        sdk::protocol::{
-            Context, CredentialType, IdentityProviderMethod, IdentityStatementRequest,
-            VerificationRequestData,
-        },
-        Web3IdAttribute,
+        // Web3IdAttribute,
     },
 };
+use concordium_rust_sdk::web3id::v1::anchor;
+use concordium_rust_sdk::web3id::v1::anchor::AnchorTransactionMetadata;
 use concordium_rust_sdk::{
     common::types::TransactionTime,
     types::WalletAccount,
     v2::{self},
-    verifiable_presentation::protocol_v1::{
-        create_and_anchor_verification_request, AnchorTransactionMetadata,
-    },
 };
 use rand::Rng;
 use std::{collections::HashMap, marker::PhantomData, path::PathBuf};
@@ -70,7 +70,7 @@ async fn main() -> anyhow::Result<()> {
     let nonce: [u8; 32] = rng.gen(); // Note: This nonce has to be generated fresh/randomly for each request.
     let connection_id = "MyWalletConnectTopic".to_string(); // Note: Use the wallet connect topic in production.
     let context_string = "MyGreateApp".to_string();
-    let context = Context::new_simple(nonce, connection_id, context_string);
+    let context = UnfilledContextInformation::new_simple(nonce, connection_id, context_string);
 
     let attribute_in_range_statement = AtomicStatement::AttributeInRange {
         statement: AttributeInRangeStatement {
@@ -82,7 +82,7 @@ async fn main() -> anyhow::Result<()> {
     };
 
     let verification_request_data = VerificationRequestData::new(context).add_statement_request(
-        IdentityStatementRequest::default()
+        RequestedIdentitySubjectClaims::default()
             .add_issuer(IdentityProviderMethod::new(0u32, Network::Testnet))
             .add_source(CredentialType::IdentityCredential)
             .add_statement(attribute_in_range_statement),
@@ -98,8 +98,8 @@ async fn main() -> anyhow::Result<()> {
         expiry,
     };
 
-    let verification_request = create_and_anchor_verification_request(
-        client.clone(),
+    let verification_request = anchor::submit_verification_request_anchor(
+        &mut client,
         anchor_transaction_metadata,
         verification_request_data,
         public_info,