+++ src/web3id.rs
@@ -1,4 +1,7 @@
-//! Functionality for retrieving, verifying, and registering web3id credentials.
+//! Functionality for retrieving, verifying, and registering Web3Id credentials.
+
+/// Functionality for retrieving, verifying, and registering V1 Web3Id credentials.
+pub mod v1;
 
 use crate::{
     cis4::{Cis4Contract, Cis4QueryError},
@@ -9,10 +12,7 @@ use concordium_base::{
     base::CredentialRegistrationID,
     cis4_types::CredentialStatus,
     contracts_common::AccountAddress,
-    id::{
-        constants::{ArCurve, IpPairing},
-        types::IpIdentity,
-    },
+    id::{constants::ArCurve, types::IpIdentity},
     web3id,
 };
 use futures::TryStreamExt;
@@ -49,7 +49,7 @@ pub struct CredentialWithMetadata {
     /// The status of the credential at a point in time.
     pub status: CredentialStatus,
     /// The extra public inputs needed for verification.
-    pub inputs: CredentialsInputs<IpPairing, ArCurve>,
+    pub inputs: CredentialsInputs<ArCurve>,
 }
 
 /// Retrieve and validate credential metadata in a particular block.
@@ -81,7 +81,7 @@ pub async fn verify_credential_metadata(
     }
     let bi = bi.into_block_identifier();
     match metadata.cred_metadata {
-        CredentialMetadata::Account(AccountCredentialMetadata { issuer, cred_id }) => {
+        CredentialMetadata::Account { issuer, cred_id } => {
             let ai = client
                 .get_account_info(&cred_id.into(), BlockIdentifier::LastFinal)
                 .await?;
@@ -139,7 +139,7 @@ pub async fn verify_credential_metadata(
                 }
             }
         }
-        CredentialMetadata::Web3Id(Web3idCredentialMetadata { contract, holder }) => {
+        CredentialMetadata::Web3Id { contract, holder } => {
             let mut contract_client = Cis4Contract::create(client, contract).await?;
             let issuer_pk = contract_client.issuer(bi).await?;