+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased changes
 
-- Added `create_and_anchor_verification_request`, `create_and_anchor_audit_record`, and `verify_and_anchor_audit_record` functions for creating and submitting anchors on-chain in the `verifiable_presentation::protocol_v1` module.
+- Added `web3id::v1::anchor` module that contains functionality to create verification request anchor (VRA) and verification audit anchor (VAA) on chain.
 
 ## 8.0.0